refactor(channel): extract slash-command dispatcher into channel_commands.go
Largest single Phase 2 extraction: the 700-LOC slash-command dispatcher
moves out of channel.go into a dedicated kin file. Kept as one big switch
intentionally - readability over cleverness; alternative (per-command
files) would scatter the shared clearCurrent / clearMain / clearThread
closures across 50+ files. Each case is independently reviewable.

  - cmd/wuphf/channel.go: -698 LOC (4249 -> 3551)
  - cmd/wuphf/channel_commands.go: +725 LOC (new)

Methods moved:
  - runActiveCommand (focus-aware wrapper)
  - runCommand (the main switch - 50+ slash commands)
  - maybeActivateChannelPickerFromInput (auto-picker on /switch typing)

renderActivePopup stays in channel.go (rendering, used by mouse hit-test).

No behavior change. All existing tests pass.

Cumulative channel.go reduction across Phase 2:
  4807 -> 4687 (pickers)        -120
       -> 4463 (sidebar)        -224
       -> 4249 (mouse)          -214
       -> 3551 (commands)       -698
  Net: -1256 LOC; channel.go shrunk by 26% so far.
